refactor: move forecast modules into forecast/ package (Phase 8 of #625)
