+++ custom_components/localshift/computation_engine_lib/__init__.py
@@ -1,5 +1,14 @@
 """Computation engine library modules."""
 
+from ..forecast.accuracy import ForecastAccuracyEngine
+from ..forecast.history import HistoryFetcher
+from ..forecast.history_store import ForecastHistoryStore
+from ..forecast.pipeline import ForecastPipeline
+from ..forecast.solar import (
+    get_solar_for_5min_slot,
+    get_solar_for_15min_slot,
+    sum_solar_before_target,
+)
 from .decision_outcome_tracker import (
     DecisionOutcomeTracker,
     DecisionRecord,
@@ -7,10 +16,6 @@ from .decision_outcome_tracker import (
 )
 from .excess_solar import ExcessSolarEngine
 from .excess_solar_signals import ExcessSolarSignalsEngine
-from .forecast_accuracy import ForecastAccuracyEngine
-from .forecast_history_store import ForecastHistoryStore
-from .forecast_pipeline import ForecastPipeline
-from .history_fetcher import HistoryFetcher
 from .optimization_controller import (
     ContextualAdjustment,
     ObjectiveWeights,
@@ -32,11 +37,6 @@ from .price_calculator import (
 from .price_signal_engine import PriceSignalEngine
 from .slot_builder import SlotBuilder, SlotBuildMetadata
 from .soc_simulator import SocSimulator
-from .solar_utils import (
-    get_solar_for_5min_slot,
-    get_solar_for_15min_slot,
-    sum_solar_before_target,
-)
 from .spike_analyzer import SpikeAnalyzer
 from .utils import (
     analyze_spike_window,
