+++ custom_components/localshift/coordinator/data.py
@@ -9,12 +9,12 @@ from typing import TYPE_CHECKING, Any
 from ..const import BatteryMode
 
 if TYPE_CHECKING:
-    from ..computation_engine_lib.forecast_accuracy import ExtendedAccuracyMetrics
+    from ..forecast.accuracy import ExtendedAccuracyMetrics
 
 
 def _default_extended_accuracy_metrics() -> Any:
     """Factory for ExtendedAccuracyMetrics that avoids circular import at module load."""
-    from ..computation_engine_lib.forecast_accuracy import ExtendedAccuracyMetrics
+    from ..forecast.accuracy import ExtendedAccuracyMetrics
 
     return ExtendedAccuracyMetrics()
