+++ custom_components/localshift/forecast/bootstrapper.py
@@ -8,8 +8,8 @@ from collections.abc import Callable
 
 from homeassistant.core import HomeAssistant
 
-from .const import CONF_SOLCAST_FORECAST_TODAY
-from .coordinator.data import CoordinatorData
+from ..const import CONF_SOLCAST_FORECAST_TODAY
+from ..coordinator.data import CoordinatorData
 
 _LOGGER = logging.getLogger(__name__)
