+++ tests/forecast/test_bootstrapper.py
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from custom_components.localshift.coordinator import CoordinatorData
-from custom_components.localshift.forecast_bootstrapper import ForecastBootstrapper
+from custom_components.localshift.forecast.bootstrapper import ForecastBootstrapper
 from tests.fixtures.ha_entities import MockState, MockStates
