+++ tests/forecast/test_history_bridging.py
