+++ tests/forecast/test_load.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from custom_components.localshift.computation_engine_lib.load_forecaster import (
+from custom_components.localshift.forecast.load import (
     LoadForecaster,
 )
 from custom_components.localshift.const import (
