+++ tests/forecast/test_solar_5min.py
@@ -7,7 +7,7 @@ instead of overlap-weighted accumulation.
 
 from datetime import datetime, timedelta, timezone
 
-from custom_components.localshift.computation_engine_lib.solar_utils import (
+from custom_components.localshift.forecast.solar import (
     get_solar_for_5min_slot,
     get_solar_for_15min_slot,
 )
